fix: refactor search form verification
